products/paid/changelog-kit/backend/pyproject.toml
[tool.poetry]
name = "changelog-kit-backend"
version = "0.1.0"
description = "Backend for Changelog Kit"
authors = ["BillMentor <billwill.mentor@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.109.0"
uvicorn = "^0.27.0"
python-frontmatter = "^1.1.0"
markdown = "^3.5.2"
pydantic-settings = "^2.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
